Split util into submodules
