+++ tests/test_util.py
@@ -5,14 +5,9 @@ import numpy as np
 import pytest
 
 from fmu.ensemble import etc
-from fmu.ensemble.util import (
-    cumcolumn_to_ratecolumn,
-    flatten,
-    normalize_dates,
-    parse_number,
-    shortcut2path,
-)
-
+from fmu.ensemble.util import flatten, parse_number, shortcut2path
+from fmu.ensemble.util.dates import normalize_dates
+from fmu.ensemble.util.rates import cumcolumn_to_ratecolumn
 
 fmux = etc.Interaction()
 logger = fmux.basiclogger(__name__, level="INFO")
